api/requirements.txt
# Core
pytz==2019.3
Django==2.2.12
django-configurations==2.2
gunicorn==20.0.4
newrelic==5.4.1.134
django-guardian==2.2.0
requests==2.23.0

# For the persistence stores
psycopg2-binary==2.8.4
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.0.0
django_unique_upload==0.2.1
django-computedfields==0.0.17

# Rest apis
djangorestframework==3.9.4
Markdown==3.1.1
django-filter==2.2.0
drf-extensions==0.6.0

# Developer Tools
ipdb==0.12.3
ipython==7.11.1
mkdocs==1.0.4
flake8==3.7.9
django-extensions==2.2.8

# Testing
mock==3.0.5
factory-boy==2.12.0
coverage==5.0.3

# Static and Media Storage
django-storages==1.8.0
boto3==1.11.1
uwsgi==2.0.18

# ElasticSearch
django-elasticsearch-dsl-drf==0.20.8
django-elasticsearch-dsl==6.4.2
elasticsearch-dsl==6.4.0
elasticsearch==6.4.0
drf_yasg
